more flexible with audio spec
The sdl3 docs lead me to believe that you always get the audio_spec you
ask for, but that is not true, and lead to some subtle crackling when
the underlying device was expecting 48k and I was giving it 44.1k.

Now we don't ask for anything in particular and try to match what we are
given.